requirements.txt
arabic-reshaper==2.1.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
elasticsearch==7.13.1
Flask==2.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
future==0.18.2
fuzzywuzzy==0.18.0
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
JSON-minify==0.3.0
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
pycparser==2.20
PyJWT==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-Levenshtein-wheels==0.13.2
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.17
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
